Initial implementation.
It works by encoding the source code of the components as a base64 strings and adding it to the id of the virtual module.
The the vite plugin decodes that base64 string when serving that virtual module. From there the regular svelte compiler takes
over and it all works.

The upside of this approach is that is very simple as it doesn't require cross-worker comunication.
The downside is that the id of the virtual modules are obnoxiously long, of the likes of:

```
virtual:inline-svelte:CiAgICAgIDxzY3JpcHQ+aW1wb3J0IERlZmluaXRpb25FbnRyeSBmcm9tICckbGliL0RlZmluaXRpb25FbnRyeS5zdmVsdGUnOzwvc2NyaXB0PgogICAgICA8RGVmaW5pdGlvbkVudHJ5IGJhY2tncm91bmQ9ImdyYXkiPgogICAgICAgIDxzdmVsdGU6ZnJhZ21lbnQgc2xvdD0iZHQiPkknbSB0aGUgZGVzY3JpcHRpb24gdGVybTwvc3ZlbHRlOmZyYWdtZW50PgogICAgICAgIDxzdmVsdGU6ZnJhZ21lbnQgc2xvdD0iZGQiPkknbSB0aGUgZGVzY3JpcHRpb24gZGVmaW5pdGlvbjwvc3ZlbHRlOmZyYWdtZW50PiAgICAKICAgICAgPC9EZWZpbml0aW9uRW50cnk+CiAgICA=.svelte
```
They can be even longer. I don't know if there might be a limit that extremely long examples might encounter, so far it doesn't seem to be. The URL spec certainly doesn't define a limit for the length of URLs.

Another downside is that it requires users repeat the script with the import statements on every test, which gets verbose and repetitive.
Some upside is that since imports live in these ad-hoc components, we don't have to worry about ESLint of typescript thinking an import statement is unused.